feat(Transitions&BattleFunctionality)
- Improved the battling script
- added a prompt to let the user know the game is starting, then Victory or defeat at the end of the battle.
- Changed the camera angle of the Battle Scene, making it more focused on team battle than pov
- Transition back to ExplorationScene
